% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands_processing.R
\name{ir_band_mean}
\alias{ir_band_mean}
\title{Compute mean price in a ratio band for use within \code{ip_band_split()}.}
\usage{
ir_band_mean(mn, mx)
}
\arguments{
\item{bands}{An array of strings containing the elements of the ratio dimension target.}
}
\value{
A single value containing the geometric mean of two values.
}
\description{
Geometric mean of each price band. For unbounded lower bounds, assumes a lower bound of 1/2 the upper bound.
For unbounded upper bounds, assumes and upper bound of 1.5x the lower bound.
}
