% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands_processing.R
\name{ir_band_split}
\alias{ir_band_split}
\title{Process array of price band strings for use in \code{ir_split_into_bands()}.}
\usage{
ir_band_split(bands)
}
\arguments{
\item{bands}{An array of strings containing the elements of the ratio dimension target.}
}
\value{
A tibble of the input strings, along with the minimum, maximum, and geometric means each one represents.
}
\description{
Splits an array of strings into minimum, maximum, and mean values for use in the \code{ir_split_into_bands()} function.
}
\examples{

bands <- c("<$200", "$200-$500", "$500+")
ir_band_split(bands)

}
