% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncnorm.R
\name{ir_dist_truncnorm_a}
\alias{ir_dist_truncnorm_a}
\title{Calculate initial denominator and numerator distributions using trunacated norm sampling. For use in \code{ir_split_into_bands()}.}
\usage{
ir_dist_truncnorm_a(n, b.mean, b.sd, df_bands)
}
\arguments{
\item{n}{number of observations.}

\item{b.mean}{mean value of distribution.}

\item{b.sd}{standard deviation of distribution, expressed as percent of mean.}

\item{df_bands}{data frame of price bands, output of \code{ir_band_split()}}
}
\value{
A data frame of each of the price bands from \code{df_bands}, as well as the probability distributions for the numerator and denominators.
}
\description{
Creates a data frame of denominator and numerator distributions for the starting point before iteration.
}
