% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_ratios.R
\name{ir_ratio_scale}
\alias{ir_ratio_scale}
\title{Scale numerator and denominator to ratios that align to bands for use in \code{ir_split_into_bands()}.}
\usage{
ir_ratio_scale(df, dims, smash_param)
}
\arguments{
\item{df}{A data frame of categorical dimensions, as well as intermediary data from \code{ir_split_into_bands()}.}

\item{smash_param}{For out-of-bound ratios, how much to increase/decrease ratio above/below min/max before next iteration.}

\item{dim}{An array of strings of the names of categorical dimensions to group data.}
}
\value{
A data frame in same shape as input, with numerator and denominator scaled to new ratio targets
}
\description{
Creates new ratios for rows of data with ratios that exceed the min/max. Recalculates the numerators and denominators, scales to targets, and performs new check.
This function is designed to be iterated over multiple times until all ratio checks pass.
}
