% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_tool.R
\name{ir_pricetool}
\alias{ir_pricetool}
\title{Wrapper function for \code{ir_split_into_bands()}, pre-populated with Price Tool inputs.}
\usage{
ir_pricetool(df, target_dim, numerator = "End-User Spending",
  denominator = "Shipments", sd = 1/3, sd_default = 1/3, sd_name = "sd",
  ratio_name = "PRICE", minimum_distribution = 0, seed_numer = NULL,
  seed_numer_wght = 0.5, seed_denom = NULL, seed_denom_wght = 0.5,
  ratio_input = NULL, ratio_input_name = "PRICE_calc",
  ratio_bounds = NULL, ratio_bounds_names = c("min", "max"),
  smash_param = 1/10, max_iterations = 40, save.intermediates = FALSE,
  save.ratio = FALSE, show.messages = TRUE)
}
\arguments{
\item{df}{A data frame of categorical dimensions and two value columns: the numerator and denominator.}

\item{sd}{Standard deviation of truncated normal distribution, expressed as a percent of mean. Can be a single number, or a data frame of values differing by element.}

\item{sd_default}{If \code{sd} is a data frame, value to use for missing elements.}

\item{sd_name}{If \code{sd} is a data frame, name of value to use as standard deviatation.}

\item{ratio_name}{Name of the new categorical dimension containing ratio bands (e.g. Price)}

\item{minimum_distribution}{Optional value of the minimum percent share of any new band element. (e.g. 0.01 means no bands can contain less than 1 percent of group total.)}

\item{seed_numer}{Optional data frame of initial starting distributions of the numerator.}

\item{seed_numer_wght}{Value between 0 and 1 of how much weight should be added to \code{seed_numer}. Larger values may not converge.}

\item{seed_denom}{Optional data frame of initial starting distributions of the denominator.}

\item{seed_denom_wght}{Value between 0 and 1 of how much weight should be added to \code{seed_denom}. Larger values may not converge.}

\item{ratio_input}{Optional data frame of average price per unit in a ratio band for all or a subset of rows or columns.}

\item{ratio_input_name}{Name of input ratio value in \code{ratio_input}.}

\item{ratio_bounds}{Optional data frame of minimum and maximum ratios for specific rows or groups. Uses \code{dplyr::left_join()}, so all values will be replicated over all elements in excluded dimensions.}

\item{ratio_bounds_names}{Names of the minimum and maximum ratio values in \code{ratio_bounds}.}

\item{smash_param}{For out-of-bound ratios, how much to increase/decrease ratio above/below min/max before next iteration.}

\item{save.intermediates}{Logical. Print out all intermediate calculations with results.}

\item{save.ratio}{Logical. Print out final calculated ratios with results.}

\item{show.messages}{Logical. Print iteration reports to console.}

\item{max_iteration}{Maximum number of iterations before ceasing iterations for unconverged model runs.}
}
\value{
A data frame with additional Price Band dimension. Total over this dimension ties to initial data
}
\description{
Creates PRICE dimension containing the min and max range of a PRICE, calculated with End-User Spending / Shipments.
}
